Inline sound_priv.h
